backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
requests
